requirements.txt
torch
torchvision
nltk
Pillow